feat: separate 'delete'
